The fitness function is defined
This custom fitness function uses the perceptron step function to check
the correctness of each individual (set of weights), i.e., whether it
classifies the data in a correct way. It uses the training data to
simulate the perceptron training algorithm, that finds, iteratively,
the best set of weights that fits best to the input data.